requirements.txt
selenium
factory_boy
behave
appdirs==1.4.0
cffi==1.9.1
cryptography==1.7.2
Django==1.10
django-appconf==1.0.2
django-compressor==2.1.1
django-debug-toolbar==1.6
django-libsass==0.7
Fabric==1.13.1
gunicorn==19.6.0
idna==2.2
libsass==0.12.3
packaging==16.8
paramiko==2.1.1
pbr==1.10.0
psycopg2==2.6.2
pyasn1==0.2.2
pip==9.0.1
pycparser==2.17
pyparsing==2.1.10
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.2
stevedore==1.20.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
virtualenvwrapper-win==1.2.1
